+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * add setting for `WP_ENVIRONMENT_TYPE` in WP 5.5+
+* refactor plugin startup a bit
 
 #### 2.5.7 / 2020-07-20
 * set disabled checked option if `WP_DISABLE_FATAL_ERROR_HANDLER` is true