+++ CHANGES.md
@@ -1,4 +1,6 @@
 #### [unreleased]
+
+#### 2.6.0 / 2020-09-01
 * add setting for `WP_ENVIRONMENT_TYPE` in WP 5.5+
 * refactor plugin startup a bit